requirements.txt
setuptools
alembic==0.9.6
aniso8601==1.3.0
click==6.7
Flask==0.12.2
Flask-Compress==1.4.0
flask_bcrypt
flask_restplus
Flask-Cors==3.0.3
Flask-HAL==1.0.4
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-WTF==0.14.2
itsdangerous==0.24
Mako==1.0.7
MarkupSafe==1.0
numpy==1.14.1
PyJWT==1.4.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
six==1.11.0
Werkzeug==0.12.2
coverage==4.4.2
requests==2.18.4
simplejson==3.12.0
workalendar==2.3.1
marshmallow==2.15.0
meinheld
cerberus
schedule
xmljson
